requirements.txt
google-generativeai
langchain
langchain-community
langchain-google-genai
faiss-cpu
python-dotenv
streamlit
fastapi
tqdm
